[ci.rb] Add attempts variable to allow CI steps to retry
Some CI steps might flake and currently that fails the whole local CI run.

To work around that, there's a new `attempts: 1` optional variable that can be set on a step to retry. By default it's 1, meaning no retries. If set to a higher number, the step will be retried up to that many times before failing. If it's not positive, there will be an ArgumentError. New test cases have been added for this feature.